+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added `mouseup`, `mousedown`, `mousemove` and implemented `mousedrag` event
 * Enabled selecting values in a `HexLogLine`
 * Added `LogLin#getValueForClipboard`, for setting the value to copy to clipboard
+* Starting Janeway with a binary file as argument will show the hex viewer
 
 ## 0.2.5 (2018-03-02)